Example: use component refs instead of recursive includes.
This is an example of the use of internal $refs for schema reuse. The
existing implementation used external includes that included schemas
recursively, and the tooling for the Eduhub gateway breaks on those
constructs.

This new setup should result in an overall smaller specification since
reuse is explicit via components.

This commit implements the changes only for EducationSpecification as
an example and to validate the technique. The idea is to implement
this for all schema objects that are used in the spec.

The same strategy can be used for responses, parameters, and
enumerations (which are a specific subset of schemas). See
https://swagger.io/docs/specification/components/ for the full list of
available component types.