Downgrade to Openapi v3.0.3 for better tooling support
Current tooling & libraries generally do not yet support openapi
3.1.0. For instance, none of the data validation libraries at
https://openapi.tools/#data-validators supports it. 3.0.3 support is
widespread.

This change sets the version to 3.0.3 and fixes an issue with the
default values for the `sort` parameter, which should be an array, so
its default should also be an array.

Further consequences: the `duration` format is not explicitly
supported in 3.0.3, so strictly conforming validators will allow any
string value for `duration` fields. This patch replaces the `duration`
format with an explicit pattern matching the ISO8601 duration format.